# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario",
        "description": "Documentación de la API del sistema de inventario",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/sucursales": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Obtener todas las sucursales del negocio autenticado",
                "operationId": "408211be2806cd0c1d2b1af59f28524d",
                "responses": {
                    "200": {
                        "description": "Lista de sucursales"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Crear una sucursal",
                "operationId": "e384af7ea38d722496dda3b0b4c661a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sucursal Central"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Principal 123"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "70000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucursal creada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sucursales/{id}": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Obtener una sucursal por ID",
                "operationId": "dc5376a7e3eed0ccd592fed477df9b42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal encontrada"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Actualizar sucursal",
                "operationId": "fdb2e84586cb6100162130e6a3fa6473",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal actualizada"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Eliminar sucursal",
                "operationId": "3a2f005ad61ae0aecbc9cc8cc11f5e25",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucursal eliminada correctamente"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sucursales/buscar": {
            "get": {
                "tags": [
                    "Sucursales"
                ],
                "summary": "Buscar sucursales por nombre",
                "operationId": "3721b9f419908c7cede47ce6e4e14f45",
                "parameters": [
                    {
                        "name": "texto",
                        "in": "query",
                        "description": "Texto a buscar en el nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado de búsqueda"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}